fix(ci): run cloud services tests in the cloud lane + schedule sleep-wake smoke
cloud-tests.yml triggers on packages/cloud/services/** but the test:cloud
runner never swept that directory, and container-control-plane/operator had
no test script, so their suites (the H4 fail-closed internal-token gate, the
#12268 annotation-corruption proof) executed in NO CI lane — a services
change went green without its tests running.

- test-cloud-run.mjs: sweep packages/cloud/services/* bun:test files with
  the same exclude + fail-loud-on-empty guards as the other roots (all 10
  services packages, 20 files — the gateway suites double-covered by their
  dedicated workflows pass in this harness too).
- container-control-plane, operator: add "test": "bun test" so the
  workspace sweep (run-all-tests.mjs) also runs them; docs updated.
- operator: extract getPreviousAgentIds into pepr-free previous-agents.ts —
  the reconciler's static pepr import breaks under bun --isolate from the
  staging dir (pepr's CJS bundle require()s kubernetes-fluent-client, whose
  top-level await makes that require throw), and the unit test only needs
  the pure parser.
- operator redis.ts: resolve ioredis-mock from the module location instead
  of process.cwd() so the MOCK_REDIS=1 test path works when bun test runs
  from the cloud lane's staging directory.
- hetzner-sleep-wake-smoke.yml: nightly cron at 06:00 UTC (an hour before
  hetzner-e2e's 07:00 slot; shared no-cancel concurrency group) with the
  honest-skip secret gating copied from hetzner-e2e.yml — missing/rejected
  HCLOUD_TOKEN_CI skips green with a loud warning on schedule/push and
  still fails a manual dispatch.

Part of #15603 (A1+A2).
