+++ packages/cloud/services/operator/capabilities/reconciler.ts
@@ -2,6 +2,7 @@
 import { K8s, Log } from "pepr";
 import { applyResources } from "./controller/generators";
 import { Server } from "./crd/generated/server-v1alpha1";
+import { getPreviousAgentIds } from "./previous-agents";
 import {
   cleanupServer,
   removeAgentServer,
@@ -102,33 +103,6 @@ export async function finalizer(instance: Server) {
   await cleanupServer(name, agentIds);
 
   Log.info(`Server ${name}: Redis cleanup complete`);
-}
-
-export function getPreviousAgentIds(instance: Server): string[] {
-  const annotation =
-    instance.metadata?.annotations?.["eliza.ai/previous-agents"];
-  // Absent annotation legitimately means "no prior agents" (first reconcile).
-  if (!annotation) return [];
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(annotation);
-  } catch (err) {
-    // A corrupt annotation must NOT read as "no previous agents": that would
-    // silently skip removeAgentServer() for agents dropped since the last
-    // reconcile, leaking stale Redis routing keys. Throw to the reconcile
-    // loop's J1 boundary (below) so the failure is logged and the annotation
-    // is not overwritten with a lie.
-    throw new Error(
-      `Server ${instance.metadata?.name ?? "<unknown>"}: corrupt eliza.ai/previous-agents annotation`,
-      { cause: err },
-    );
-  }
-  if (!Array.isArray(parsed) || parsed.some((id) => typeof id !== "string")) {
-    throw new Error(
-      `Server ${instance.metadata?.name ?? "<unknown>"}: eliza.ai/previous-agents annotation is not a string array`,
-    );
-  }
-  return parsed;
 }
 
 async function updateStatus(instance: Server, status: Server["status"]) {
